requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.12.1
django-cleanup==5.1.0
django-countries==6.1.3
django-datetime-widget==0.9.3
django-dynamic-formsets==0.0.8
django-jquery==3.1.0
mock==4.0.2
munch==2.5.0
Pillow==7.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.1
pyuca==1.2
six==1.15.0
sqlparse==0.3.1